Fix current history not shown in history multiview
To reproduce the problem go to the saved histories menu, select a
history that is not the most recently updated (i.e not the first
history). Note which history *is* the most recently updated. If you go
to the multiple history view, the most recently updated history doesn't
show (it does appear if you search for it though).
This commit removes parts of 2a38b11 and d89843e, which was first
release in 16.10, at which point the problem appears to have first
occured
(https://github.com/galaxyproject/galaxy/issues/3519#issuecomment-280659270).